- fix errors after renaming
