+++ src/components/keyshortcuts/generateHotkeys.js
